Add LocationUpdater
